refactor: refactor presenter mode
split up into separate files, basic housekeeping, keep working index.js with prefix
