+++ @dekk/deck/package.json
@@ -13,6 +13,7 @@
     "posttest": "yarn lint"
   },
   "dependencies": {
+    "@dekk/countdown": "^1.0.0-5",
     "@dekk/slide": "^1.0.0-5",
     "@dekk/store": "^1.0.0-5",
     "@dekk/utils": "^1.0.0-5",
